Added nodemon and initial project structure
