+++ src/app.ts
@@ -1,0 +1,11 @@
+import express from "express";
+import workflowsRouter from './routers/workflowsRouter'
+const app = express();
+//TODO: Move midlewares setup to separate files under middlewares folder
+express.json();
+
+//Routers config
+app.use('/api/workflows', workflowsRouter);
+
+// Export express instance
+export default app;
